Move MFC supporting functions to minimfc
